Finished working on the take lesson pathway.
Quiz screen displays the correct text and image from file. Tracks if the correct answer is given and updates the current score for the lesson.

Lesson complete screen updates the completion for the user, checks how the score compared to the users current best and gives a message reflecting how it compares, updating the users recorded score appropriately.