Added Dictionary layout and Java file.
Added reference to Dictionary in Android Manifest.

Added launch intent to Menu, to launch the Dictionary activity.

Added necessary strings for Dictionary.

NOTE: Dictionary is only a demo, it does not have an actual search function. User is meant to choose two different languages and press 'search', which will then produce a translation for demonstration purposes.